+++ builder/AbstractBuilder.ts
@@ -149,9 +149,11 @@ class AbstractBuilder extends jtree.TreeNode {
       equal(exampleErrors.length, 0, exampleErrors.length ? "examples errs: " + exampleErrors : "no example errors")
     }
 
-    testTree[`grammarTypeScriptImplementationCheckOf${grammarPath}`] = (equal: Function) => checkGrammarFile(equal, new jtree.GrammarProgram(Disk.read(grammarPath)))
+    const grammarProgram = new jtree.HandGrammarProgram(Disk.read(grammarPath))
+
+    testTree[`grammarTypeScriptImplementationCheckOf${grammarPath}`] = (equal: Function) => checkGrammarFile(equal, grammarProgram)
     testTree[`grammarGrammarImplementationCheckOf${grammarPath}`] = (equal: Function) => checkGrammarFile(equal, jtree.makeProgram(grammarPath, __dirname + "/../langs/grammar/grammar.grammar"))
-    testTree[`examplesInGrammarCheckForGrammarTypeScriptImplementationCheckOf${grammarPath}`] = (equal: Function) => checkGrammarExamples(equal, new jtree.GrammarProgram(Disk.read(grammarPath)))
+    testTree[`examplesInGrammarCheckForGrammarTypeScriptImplementationCheckOf${grammarPath}`] = (equal: Function) => checkGrammarExamples(equal, grammarProgram)
 
     return testTree
   }