+++ coreTests/GrammarLanguage.compiled.test.ts
@@ -17,7 +17,7 @@ const mkdirp = require("mkdirp")
 mkdirp.sync(outputDir)
 
 const makeProgram = (grammarCode: string, code: string) => {
-  const grammarProgram = new jtree.GrammarProgram(grammarCode)
+  const grammarProgram = new jtree.HandGrammarProgram(grammarCode)
   const rootProgramConstructor = grammarProgram.compileAndReturnRootConstructor()
   return new rootProgramConstructor(code)
 }