+++ coreTests/GrammarLanguage.test.ts
@@ -7,7 +7,7 @@ import { treeNotationTypes } from "../products/treeNotationTypes"
 
 const { Disk } = require("../products/Disk.node.js")
 
-const GrammarProgram = jtree.GrammarProgram
+const HandGrammarProgram = jtree.HandGrammarProgram
 const jibberishRootDir = __dirname + "/../langs/jibberish/"
 
 const numbersPath = __dirname + "/../langs/numbers/numbers.grammar"
@@ -24,7 +24,7 @@ const testTree: treeNotationTypes.testTree = {}
 
 testTree.emptyProgram = equal => {
   // Arrange/Act/Assert
-  const program = new GrammarProgram()
+  const program = new HandGrammarProgram()
   const errs = program.getAllErrors()
 
   // Assert
@@ -34,7 +34,7 @@ testTree.emptyProgram = equal => {
 
 testTree.grammarLangBasics = equal => {
   // Arrange/Act
-  const grammarProgram = new GrammarProgram(jibberishGrammarCode)
+  const grammarProgram = new HandGrammarProgram(jibberishGrammarCode)
   const errs = grammarProgram.getAllErrors()
 
   // Assert
@@ -59,7 +59,7 @@ const makeIrisProgram = (code: string) => makeProgram(Disk.read(__dirname + "/..
 const makeNumbersProgram = (code: string) => makeProgram(numbersGrammar, code)
 
 const makeProgram = (grammarCode: string, code: string) => {
-  const grammarProgram = new GrammarProgram(grammarCode)
+  const grammarProgram = new HandGrammarProgram(grammarCode)
   const rootProgramConstructor = grammarProgram.compileAndReturnRootConstructor()
   return new rootProgramConstructor(code)
 }
@@ -154,7 +154,7 @@ const langs = Disk.dir(__dirname + `/../langs/`)
 langs.forEach((lang: string) => {
   testTree[`${lang}SimTest`] = equal => {
     const grammarCode = Disk.read(__dirname + `/../langs/${lang}/${lang}.grammar`)
-    const grammarProgram = new jtree.GrammarProgram(grammarCode)
+    const grammarProgram = new jtree.HandGrammarProgram(grammarCode)
     const programConstructor = grammarProgram.compileAndReturnRootConstructor()
 
     // Act
@@ -463,7 +463,7 @@ testTree.blobNodes = equal => {
 
 testTree.sublimeSyntaxFile = equal => {
   // Arrange/Act
-  const grammarProgram = new GrammarProgram(jibberishGrammarCode)
+  const grammarProgram = new HandGrammarProgram(jibberishGrammarCode)
   const code = grammarProgram.toSublimeSyntaxFile()
 
   // Assert
@@ -472,7 +472,7 @@ testTree.sublimeSyntaxFile = equal => {
 
 testTree.toStumpString = equal => {
   // Arrange/Act
-  const grammarProgram = new GrammarProgram(arrowGrammar).compileAndReturnRootConstructor()
+  const grammarProgram = new HandGrammarProgram(arrowGrammar).compileAndReturnRootConstructor()
   const code = new grammarProgram()
     .getDefinition()
     .getNodeTypeDefinitionByNodeTypeId("chargeNode")
@@ -525,7 +525,7 @@ div
 
 testTree.minimumGrammar = equal => {
   // Arrange/Act
-  const programConstructor = new GrammarProgram(
+  const programConstructor = new HandGrammarProgram(
     `anyLangNode
  root
  catchAllNodeType anyNode
@@ -543,14 +543,14 @@ anyCell`
   equal(errors.length, 0)
 
   // Arrange/Act/Assert
-  const constructor = new GrammarProgram().compileAndReturnRootConstructor()
+  const constructor = new HandGrammarProgram().compileAndReturnRootConstructor()
   const errs = new constructor("foobar").getAllErrors()
   equal(errs.length, 0)
 }
 
 testTree.rootCatchAllNode = equal => {
   // Arrange
-  const abcLang = new GrammarProgram(`abcNode
+  const abcLang = new HandGrammarProgram(`abcNode
  root`).compileAndReturnRootConstructor()
 
   // Act/Assert
@@ -559,7 +559,7 @@ testTree.rootCatchAllNode = equal => {
   equal(program.toCellTypeTree(), "extraWordCell", "one word")
 
   // Arrange
-  const abcLangWithErrors = new GrammarProgram(`abcNode
+  const abcLangWithErrors = new HandGrammarProgram(`abcNode
  root
  catchAllNodeType errorNode
 errorNode
@@ -571,7 +571,7 @@ errorNode
 
 testTree.blankNodeId = equal => {
   // Arrange
-  const abcLang = new GrammarProgram(`nodeType `).compileAndReturnRootConstructor()
+  const abcLang = new HandGrammarProgram(`nodeType `).compileAndReturnRootConstructor()
 
   // Act/Assert
   equal(new abcLang("foobar").getAllErrors().length, 0)
@@ -580,7 +580,7 @@ testTree.blankNodeId = equal => {
 testTree.grammarWithLoop = equal => {
   // Arrange/Act/Assert
   try {
-    const programConstructor = new GrammarProgram(
+    const programConstructor = new HandGrammarProgram(
       `langWithLoopNode
  root
  catchAllNodeType nodeANode
@@ -641,7 +641,7 @@ foobarCell
 
 testTree.toNodeJsJavascript = equal => {
   // Arrange
-  let program = new GrammarProgram(grammarGrammar)
+  let program = new HandGrammarProgram(grammarGrammar)
   // Act
   let compiledParser = program.toNodeJsJavascript()
   // Assert
@@ -650,7 +650,7 @@ testTree.toNodeJsJavascript = equal => {
 
 testTree.invalidGrammarRegression = equal => {
   // Arrange
-  let program = new GrammarProgram(`oldStyle something
+  let program = new HandGrammarProgram(`oldStyle something
  root`)
   // Act
   let compiledParser = program.toNodeJsJavascript()
@@ -660,7 +660,7 @@ testTree.invalidGrammarRegression = equal => {
 
 testTree.bundler = equal => {
   // Arrange
-  const jibberishGrammarProgram = new GrammarProgram(jibberishGrammarCode)
+  const jibberishGrammarProgram = new HandGrammarProgram(jibberishGrammarCode)
 
   // Act
   const bundle = jibberishGrammarProgram.toBundle()
@@ -671,14 +671,14 @@ testTree.bundler = equal => {
 
 testTree.examples = equal => {
   // Arrange/Act
-  const jibberishGrammarProgram = new GrammarProgram(jibberishGrammarCode)
+  const jibberishGrammarProgram = new HandGrammarProgram(jibberishGrammarCode)
 
   // Assert
   let errors = jibberishGrammarProgram.getErrorsInGrammarExamples()
   equal(errors.length, 0)
 
   // Arrange/Act
-  const badGrammarProgram = new GrammarProgram(
+  const badGrammarProgram = new HandGrammarProgram(
     `badNode
  root
  inScope addNode