+++ coreTests/UnknownGrammarProgram.test.ts
@@ -46,7 +46,7 @@ langs.forEach((name: string) => {
 
     // Act
     const inferredPrefixGrammarCode = new UnknownGrammarProgram(sampleCode).inferGrammarFileForAKeywordLanguage("foobar")
-    const inferredPrefixGrammarProgram = new jtree.GrammarProgram(inferredPrefixGrammarCode)
+    const inferredPrefixGrammarProgram = new jtree.HandGrammarProgram(inferredPrefixGrammarCode)
     const rootProgramConstructor = inferredPrefixGrammarProgram.compileAndReturnRootConstructor()
     const programParsedWithInferredGrammar = new rootProgramConstructor(sampleCode)