+++ langs/grammar/grammar.node.js
@@ -1,1 +1,1 @@
-module.exports = require("../../index.js").jtree.getProgramConstructor(__dirname + "/grammar.grammar")
+module.exports = require("../../index.js").jtree.compileGrammarFileAtPathAndReturnRootConstructor(__dirname + "/grammar.grammar")