+++ products/dumbdown.browser.js
@@ -22,7 +22,7 @@
     }
     getGrammarProgram() {
       if (!this._cachedGrammarProgramRoot)
-        this._cachedGrammarProgramRoot = new jtree.GrammarProgram(`anyCell
+        this._cachedGrammarProgramRoot = new jtree.HandGrammarProgram(`anyCell
 blankCell
 dashCell
  highlightScope constant.language