+++ products/hakon.browser.js
@@ -18,7 +18,7 @@
     }
     getGrammarProgram() {
       if (!this._cachedGrammarProgramRoot)
-        this._cachedGrammarProgramRoot = new jtree.GrammarProgram(`anyCell
+        this._cachedGrammarProgramRoot = new jtree.HandGrammarProgram(`anyCell
 keywordCell
 commentKeywordCell
  extends keywordCell