+++ products/hakon.nodejs.js
@@ -21,7 +21,7 @@
     }
     getGrammarProgram() {
       if (!this._cachedGrammarProgramRoot)
-        this._cachedGrammarProgramRoot = new jtree.GrammarProgram(`anyCell
+        this._cachedGrammarProgramRoot = new jtree.HandGrammarProgram(`anyCell
 keywordCell
 commentKeywordCell
  extends keywordCell